Moved commonlib/lt.py into hazardlib
